NIFI-6363 (local) Moved CipherUtils class.
