+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-properties-loader/src/test/java/org/apache/nifi/properties/sensitive/keystore/KeyStoreWrappedSensitivePropertyProviderIT.java
@@ -20,7 +20,7 @@ import org.apache.nifi.properties.sensitive.AbstractSensitivePropertyProviderTes
 import org.apache.nifi.properties.sensitive.ByteArrayKeyStoreProvider;
 import org.apache.nifi.properties.sensitive.SensitivePropertyConfigurationException;
 import org.apache.nifi.properties.sensitive.SensitivePropertyProvider;
-import org.apache.nifi.security.util.CipherUtils;
+import org.apache.nifi.properties.sensitive.CipherUtils;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.junit.Assert;
 import org.junit.Before;
